+++ CHANGELOG.md
@@ -11,7 +11,7 @@ about writing changes to this log.
 ## [Unreleased]
 
 - [TVIST1-729](https://jira.itkdev.dk/browse/TVIST1-729):
-  Allow hearing to start with any one part.
+  Allowed hearing to start with just a party or counterparty.
 
 ## [1.0.3] 2022-12-19